baitap

+++ module4/ss12_jquery_and_ajax/practice/phone/src/main/java/com/example/repository/ISmartphoneRepository.java
@@ -3,5 +3,5 @@ package com.example.repository;
 import com.example.model.Smartphone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ISmartphoneRepository extends JpaRepository<Smartphone,Integer> {
+public interface ISmartphoneRepository extends JpaRepository<Smartphone,Long> {
 }
